Use k8s Service to create fake avalanche targets for Prometheus
Currently we use k8s StatefulSet and native replica to create multiple
Avalanche instances to generate series load to Prometheus. However this is
less optimal and occupies real CPU & memory in a benchmarking environment.

This fix creates fake Service to mimic existence of multiple avalanche
with less demanding resources.

A new helm value `fakePrometheusTargetCount` has been introduced to declare the
number of fake prometheus targets to be created for avalanche.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>